Fix Linux GCC build warnings and test failures
- Fix variable shadowing warning in callback_job.cpp
  - Rename lambda parameter from 'data' to 'callback_data' to avoid shadowing

- Fix sign conversion warnings in queue statistics
  - Cast both operands to double in division operations
  - Applied to lockfree_mpmc_queue.h and adaptive_job_queue.h/cpp

- Fix memory safety issue in SimpleMPMCPerformance test
  - Replace raw reference capture with shared_ptr for atomic counter
  - Prevents use-after-free when lambda outlives test scope
  - Add null check for dequeue result

- Restore proper CompilerChecks.cmake configuration
  - Re-add MinGW-specific warning adjustments
  - Keep PCH disabled for MinGW
  - Remove project header from PCH (causes issues)

These changes ensure clean compilation on Linux GCC while maintaining
compatibility with other platforms.

+++ sources/thread_base/lockfree/queues/adaptive_job_queue.cpp
@@ -279,8 +279,8 @@ namespace thread_module
 		auto ops = metrics_.operation_count.load(std::memory_order_relaxed);
 		if (ops == 0) return false;
 		
-		auto avg_latency = static_cast<double>(metrics_.total_latency_ns.load(std::memory_order_relaxed)) / ops;
-		auto contention_ratio = static_cast<double>(metrics_.contention_count.load(std::memory_order_relaxed)) / ops;
+		auto avg_latency = static_cast<double>(metrics_.total_latency_ns.load(std::memory_order_relaxed)) / static_cast<double>(ops);
+		auto contention_ratio = static_cast<double>(metrics_.contention_count.load(std::memory_order_relaxed)) / static_cast<double>(ops);
 		
 		return (contention_ratio > CONTENTION_THRESHOLD_HIGH && avg_latency > LATENCY_THRESHOLD_HIGH_NS);
 	}
@@ -290,8 +290,8 @@ namespace thread_module
 		auto ops = metrics_.operation_count.load(std::memory_order_relaxed);
 		if (ops == 0) return false;
 		
-		auto avg_latency = static_cast<double>(metrics_.total_latency_ns.load(std::memory_order_relaxed)) / ops;
-		auto contention_ratio = static_cast<double>(metrics_.contention_count.load(std::memory_order_relaxed)) / ops;
+		auto avg_latency = static_cast<double>(metrics_.total_latency_ns.load(std::memory_order_relaxed)) / static_cast<double>(ops);
+		auto contention_ratio = static_cast<double>(metrics_.contention_count.load(std::memory_order_relaxed)) / static_cast<double>(ops);
 		
 		// Switch back if contention is low and lock-free overhead isn't worth it
 		return (contention_ratio < CONTENTION_THRESHOLD_LOW && avg_latency > LATENCY_THRESHOLD_LOW_NS * 2);
